app/requirements.txt
streamlit==1.39.0
tensorflow==2.16.1
numpy==1.26.4
pillow==10.3.0
